self-supervised-learning/requirements.txt
easydict
matplotlib
numpy
opencv-python
pandas
Pillow
scikit-image
scikit-learn
SimpleITK
tensorboard
tensorboard-plugin-wit
torch
torchaudio
torchsampler
torchsummaryX
torchvision
tqdm